Load 802.15.4 transceiver's module at boot
Load at86rf230 kernel module at the system boot in order to avoid using
bash script or modprobe command to load 802.15.4 transceiver.

Note that this module request usage of device tree in order to work
properly. In Buildroot, this can be done by setting
BR2_PACKAGE_RPI_FIRMWARE_INSTALL_DTB_OVERLAYS (in Raspberry Pi 1 and 2
package is not set by default).

A guide upon how to load this module can be found in:

http://openlabs.co/blog/archives/1-6LoWPAN-kernel-on-a-Raspberry-Pi