Fix knot user permissions to run reboot applet
The busybox configuration was allowing any user to run its applets with
root privileges, including the `su` command, which is a huge security
flaw. This occurred because the FEATURE_SUID was disabled and it is
responsible for dropping all the privileges for applets that don't need
root access. The current patch solves this issue by enabling it and
configuring the applet permission via the FEATURE_SUID_CONFIG option.
This permission was changed to allow root and any user inside 'knot'
group to run the reboot applet.

Moreover, considering that isn't good to customize the configuration
by change the package files directly, the change was moved to the knot
config files.

SUID reference: http://www.softforge.de/bb/suid.html