Update rabbitmq folder ownership to rabbitmq user
After fixed the security issue that allowed any user to run applets as
root user (fc2de2e), the rabbitmq server could not start because it
was trying to write log files in a directory owned by the root user.
That is, the rabbitmq server is now running really as rabbitmq user.

In order to solve this issue, we need to change the ownership of the
rabbitmq directory in a way that the service can create what log files
it needs.