package: postgresql: Build source with plugins
PostgreSQL comes with some additional modules that can be found in
the `contrib` directory. Some of these modules are plugin features
that are not part of the core PostgreSQL system. When building from
the sources, these `contrib` modules are not built automatically,
unless if it is built with the `world` target.

As some of these plugins (pgcrypto and ltree) are needed for the
Mainflux services, this patch overrides the `make` macros of the
postgresql build process to include the `world` target.

Ref: https://www.postgresql.org/docs/10/contrib.html